11: Se agrego el controlador de registro y se movio la logica a su respectivo controlador
