+++ IntegracionBancaria/Model/Data/Dapper/PerfilDao.cs
@@ -1,0 +1,7 @@
+namespace IntegracionBancaria.Model.Data
+{
+    public class PerfilDao
+    {
+        
+    }
+}
