+++ IntegracionBancaria/Service/ServicioBanco.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using IntegracionBancaria.Model.Data.Dapper;
 using IntegracionBancaria.Model.Domain;
+using IntegracionBancaria.Model.Views;
 using Microsoft.Extensions.Logging;
 
 namespace IntegracionBancaria.Service
@@ -20,13 +21,27 @@ namespace IntegracionBancaria.Service
         
         public Result<Exception, IList<Banco>> ObtenerBancosActivos()
         {
-            try {
+            try 
+            {
+                _logger.LogInformation("Obteniendo listado de bancos activos");
                 var bancos = _bancoDao.ObtenerBancosActivos();
                 return Result<Exception, IList<Banco>>.ForSuccess(bancos);
             }
             catch(Exception ex)
             {
                 return Result<Exception, IList<Banco>>.ForFailure(ex);
+            }
+        }
+
+        public Result<Exception, long> RegistrarUsuario(Registro registro)
+        {
+            try
+            {
+                return null;
+            }
+            catch (Exception ex)
+            {
+                return Result<Exception, long>.ForFailure(ex);
             }
         }
     }
