+++ IntegracionBancaria/Views/Login/Login.cshtml
@@ -23,7 +23,7 @@
                     <input type="submit" class="btn btn-primary" value="Ingresar" />
                     <p>
                         </br>
-                        ¿No tiene usuario?&nbsp;<a asp-controller="Login" asp-action="Registro">Registrese</a>
+                        ¿No tiene usuario?&nbsp;<a asp-controller="Registro" asp-action="Registro">Registrese</a>
                     </p>
                 </form>
