+++ IntegracionBancaria/Views/Registro/Registro.cshtml
@@ -10,7 +10,7 @@
 <div class="row">
     <div class="col-md-offset-2 col-md-6">
 
-        @using (Html.BeginForm("Registrarse", "Login", FormMethod.Post, new { @class = "contacto" }))
+        @using (Html.BeginForm("Registrarse", "Registro", FormMethod.Post, new { @class = "contacto" }))
         {
             @Html.AntiForgeryToken()
