+++ Migraciones/sql/V2__11-agregar-bancos-iniciales.sql
@@ -1,0 +1,6 @@
+INSERT INTO bancos.banco (codigo,nombre,pais, activo)
+VALUES ('BR','Banco Rural','GT', true) ;
+
+INSERT INTO bancos.banco (codigo,nombre,pais, activo)
+VALUES ('BI','Banco de la Industria','GT', true) ;
+
